+++ src/main/java/muzusi/application/stockchart/scheduler/StockChartScheduler.java
@@ -1,27 +1,17 @@
 package muzusi.application.stockchart.scheduler;
 
 import lombok.RequiredArgsConstructor;
-import muzusi.application.stockchart.service.StockChartUpdater;
+import muzusi.application.stockchart.service.StockChartCreationProcessor;
 import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.scheduling.annotation.Schedules;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
 public class StockChartScheduler {
-    private final StockChartUpdater stockChartUpdater;
+    private final StockChartCreationProcessor stockChartCreationProcessor;
 
-    @Schedules({
-            @Scheduled(cron = "0 10,20,30,40,50 9 * * 1-5"),
-            @Scheduled(cron = "0 0/10 10-14 * * 1-5"),
-            @Scheduled(cron = "0 0,10,20,30 15 * * 1-5")
-    })
-    public void runUpdateStockMinutesChartJob() throws InterruptedException {
-        stockChartUpdater.updateStockMinutesChart();
-    }
-
-    @Scheduled(cron = "0 0 16 * * 1-5")
-    public void runFlushStockMinutesChartJob() {
-        stockChartUpdater.flushStockMinutesChart();
+    @Scheduled(cron = "0 0 9 * * 1-5")
+    public void runCreateStockChartJob() {
+        stockChartCreationProcessor.createStockChart();
     }
 }
