+++ src/main/java/muzusi/domain/stock/repository/StockDailyRepository.java
@@ -3,8 +3,11 @@ package muzusi.domain.stock.repository;
 import muzusi.domain.stock.entity.StockDaily;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 public interface StockDailyRepository extends MongoRepository<StockDaily, String> {
     List<StockDaily> findByStockCodeOrderByDateAsc(String stockCode);
+
+    boolean existsByDate(LocalDateTime date);
 }
